4th commit

+++ utils/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: megrisse <megrisse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 03:39:27 by megrisse          #+#    #+#             */
-/*   Updated: 2022/05/19 21:58:37 by megrisse         ###   ########.fr       */
+/*   Updated: 2022/06/18 03:44:01 by megrisse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	unsigned char	*ptr;
 	int				x;
