+++ utils/ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: megrisse <megrisse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/12 08:28:36 by megrisse          #+#    #+#             */
-/*   Updated: 2022/05/19 21:58:45 by megrisse         ###   ########.fr       */
+/*   Updated: 2022/06/18 03:52:47 by megrisse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-void	*ft_calloc(size_t count, size_t size)
+void	*ft_calloc(size_t count, int size)
 {
 	void	*ptr;
